pdf-extractor/requirements.txt
PyMuPDF==1.23.26
flask==3.0.0
werkzeug==3.0.1
gunicorn==21.2.0
python-dotenv==1.0.0
